pyproject.toml
[tool.poetry]
name = "aws-scripts"
version = "0.1.0"
description = ""
authors = ["Davi Wesley <davi.wesley@icloud.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
boto3 = "^1.26.32"

[tool.poetry.dev-dependencies]
black = "^22.12.0"
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
